soft delete/restore added

+++ src/main/java/com/ankit/assignmentspringboot/controller/UserAddressController.java
@@ -5,26 +5,42 @@ import com.ankit.assignmentspringboot.requestDto.UpdateUserAddressRequestDto;
 import com.ankit.assignmentspringboot.responseDto.GetUserAddressResponseDto;
 import com.ankit.assignmentspringboot.service.UserAddressService;
 import com.ankit.assignmentspringboot.utility.ApiResponse;
+import com.ankit.assignmentspringboot.utility.GetAuthUserId;
+import com.ankit.assignmentspringboot.utility.UserRole;
+import com.ankit.assignmentspringboot.utility.security.GetAuthUserRole;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+import java.util.Objects;
+
 @RestController
 @RequestMapping("/address")
 public class UserAddressController {
 
+    private static final Logger log = LoggerFactory.getLogger(UserAddressController.class);
     private final UserAddressService userAddressService;
+    private final GetAuthUserRole getAuthUserRole;
 
     @Autowired
-    public UserAddressController(UserAddressService userAddressService) {
-        System.out.println("inside user address controller constructor");
+    public UserAddressController(UserAddressService userAddressService, GetAuthUserRole getAuthUserRole) {
+        this.getAuthUserRole = getAuthUserRole;
+        log.info("inside user address controller constructor");
         this.userAddressService = userAddressService;
     }
 
     @PostMapping
     public ResponseEntity<ApiResponse<Void>> saveUserAddress(@RequestBody SaveUserAddressRequestDto userAddress){
         try {
+            if (
+                    !(Objects.equals(userAddress.getUser_id(), GetAuthUserId.getUserId())) &&
+                    !(List.of(UserRole.ADMIN, UserRole.MODERATOR).contains(getAuthUserRole.getUserRole()))){
+                throw new RuntimeException("operation is not allowed");
+            }
             userAddressService.saveUserAddress(userAddress);
             ApiResponse<Void> resp = new ApiResponse<>(true, "saved user address");
             return ResponseEntity.status(HttpStatus.CREATED).body(resp);
@@ -35,8 +51,13 @@ public class UserAddressController {
     }
 
     @GetMapping
-    public ResponseEntity<ApiResponse<?>> getUserAddressById(@RequestParam int id){
+    public ResponseEntity<ApiResponse<?>> getUserAddressById(@RequestParam Integer id){
         try {
+            if (
+                    !(Objects.equals(id.toString(), GetAuthUserId.getUserId())) &&
+                            !(List.of(UserRole.ADMIN, UserRole.MODERATOR).contains(getAuthUserRole.getUserRole()))){
+                throw new RuntimeException("operation is not allowed");
+            }
             GetUserAddressResponseDto existingUserAddress = userAddressService.getUserAddressById(id);
             ApiResponse<GetUserAddressResponseDto> resp = new ApiResponse<>(true, "user address for id: " + id, existingUserAddress);
             return ResponseEntity.status(HttpStatus.OK).body(resp);
@@ -49,6 +70,11 @@ public class UserAddressController {
     @PutMapping
     public ResponseEntity<ApiResponse<Void>> updateUserAddress(@RequestBody UpdateUserAddressRequestDto userAddressRequestDto){
         try {
+            if (
+                    !(Objects.equals(userAddressRequestDto.getUser_id(), GetAuthUserId.getUserId())) &&
+                            !(List.of(UserRole.ADMIN, UserRole.MODERATOR).contains(getAuthUserRole.getUserRole()))){
+                throw new RuntimeException("operation is not allowed");
+            }
             userAddressService.updateUserAddressData(userAddressRequestDto);
             ApiResponse<Void> resp = new ApiResponse<>(true, "updated user address");
             return ResponseEntity.ok(resp);
@@ -62,6 +88,11 @@ public class UserAddressController {
     public ResponseEntity<ApiResponse<Void>> deleteUserAddress(@RequestParam Integer id){
         try {
             if (id == null) throw new NullPointerException();
+            if (
+                    !(Objects.equals(id.toString(), GetAuthUserId.getUserId())) &&
+                            !(List.of(UserRole.ADMIN, UserRole.MODERATOR).contains(getAuthUserRole.getUserRole()))){
+                throw new RuntimeException("operation is not allowed");
+            }
             userAddressService.deleteUserAddress(id);
             ApiResponse<Void> resp = new ApiResponse<>(true, "user address deleted");
             return ResponseEntity.ok(resp);
